+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "week1",
+  "version": "1.0.0",
+  "description": "",
+  "main": "todo.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/DongjaJ/fe-console-todo.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/DongjaJ/fe-console-todo/issues"
+  },
+  "homepage": "https://github.com/DongjaJ/fe-console-todo#readme"
+}
